// src/commands/domain/enums/updatePlaylistButtonsEnum.ts
export const enum UpdatePlaylistButtonsEnum {
    NAME = 'change pl name',
    ADD = 'add songs to pl',
    ADDPLAYING = 'add playing songs to pl',
    REMOVE = 'remove songs from pl',
    SHOWPL = 'show songs from pl',
    SHOWADDED = 'show songs from added list',
    SHOWREMOVED = 'show songs from remove list',
    CANCEL = 'cancel create pl',
    SAVE = 'save new pl',
}